test(image): add test to image
